refactor!: move pkg/checker to internal/servicecheck
