+++ internal/servicecheck/cache.go
@@ -1,4 +1,4 @@
-package checker
+package servicecheck
 
 import "time"
